refactor: Remove dead code
