+++ core/model/src/main/kotlin/app/pachli/core/model/Suggestion.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Pachli Association
+ * Copyright 2025 Pachli Association
  *
  * This file is a part of Pachli.
  *
@@ -17,69 +17,32 @@
 
 package app.pachli.core.model
 
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
 /**
- * Sources for suggestions returned from Mastodon 3.x.
+ * Sources of suggestions for followed accounts.
  *
- * See [https://docs.joinmastodon.org/entities/Suggestion/](https://docs.joinmastodon.org/entities/Suggestion/).
- */
-enum class SuggestionSource {
-    /** "This account was manually recommended by your administration team" */
-    @Json(name = "staff")
-    STAFF,
-
-    /** "You have interacted with this account previously" */
-    @Json(name = "past_interactions")
-    PAST_INTERACTIONS,
-
-    /** "This account has many reblogs, favourites, and active local followers within the last 30 days" */
-    @Json(name = "global")
-    GLOBAL,
-
-    UNKNOWN,
-}
-
-/**
- * Sources for suggestions returned from Mastodon 4.x
- * ([https://github.com/mastodon/documentation/issues/1398](https://github.com/mastodon/documentation/issues/1398))
+ * Wraps and merges the different sources supplied across different Mastodon API
+ * versions.
  */
 enum class SuggestionSources {
     /** "Hand-picked by the {domain} team" */
-    @Json(name = "featured")
     FEATURED,
 
     /** "This profile is one of the most followed on {domain}." */
-    @Json(name = "most_followed")
     MOST_FOLLOWED,
 
     /** "This profile has been recently getting a lot of attention on {domain}." */
-    @Json(name = "most_interactions")
     MOST_INTERACTIONS,
 
     /** "This profile is similar to the profiles you have most recently followed." */
-    @Json(name = "similar_to_recently_followed")
     SIMILAR_TO_RECENTLY_FOLLOWED,
 
     /** "Popular among people you follow" */
-    @Json(name = "friends_of_friends")
     FRIENDS_OF_FRIENDS,
 
     UNKNOWN,
 }
 
-/**
- * A suggested account to follow.
- *
- * @property source The single reason for this suggestion.
- * @property sources One or more reasons for this suggestion.
- * @property account The suggested account.
- */
-@JsonClass(generateAdapter = true)
 data class Suggestion(
-    @Deprecated("Mastodon 4.x switched to sources")
-    val source: SuggestionSource,
-    val sources: List<SuggestionSources>? = null,
+    val sources: List<SuggestionSources> = emptyList(),
     val account: Account,
 )
