+++ core/network/src/main/kotlin/app/pachli/core/network/model/Attachment.kt
@@ -105,8 +105,6 @@ data class Attachment(
         val x: Float = 0f,
         val y: Float = 0f,
     ) {
-        fun toMastodonApiString(): String = "$x,$y"
-
         fun asModel() = app.pachli.core.model.Attachment.Focus(
             x = x,
             y = y,
@@ -138,15 +136,6 @@ data class Attachment(
             height = height,
             aspect = aspect,
         )
-    }
-
-    /**
-     * @return True if this attachment can be previewed. A previewable attachment
-     *     must be a known type and have a non-null width for the preview image.
-     */
-    fun isPreviewable(): Boolean {
-        if (type == Type.UNKNOWN) return false
-        return !(meta?.original?.width == null && meta?.small?.width == null)
     }
 }
