+++ core/network/src/main/kotlin/app/pachli/core/network/model/Poll.kt
@@ -4,8 +4,6 @@ import app.pachli.core.network.json.BooleanIfNull
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import java.util.Date
-import kotlin.collections.map
-import kotlin.collections.mapIndexed
 
 @JsonClass(generateAdapter = true)
 data class Poll(
@@ -21,38 +19,6 @@ data class Poll(
     @BooleanIfNull(false) val voted: Boolean = false,
     @Json(name = "own_votes") val ownVotes: List<Int>?,
 ) {
-
-    /**
-     * @param choices Indices of the user's choices
-     * @return A copy of the poll with the vote counts for each choice
-     * updated to reflect the user's voting choices in [choices], and
-     * with [voted][Poll.voted] set to `true`.
-     */
-    fun votedCopy(choices: List<Int>): Poll {
-        val newOptions = options.mapIndexed { index, option ->
-            if (choices.contains(index)) {
-                option.copy(votesCount = option.votesCount + 1)
-            } else {
-                option
-            }
-        }
-
-        return copy(
-            options = newOptions,
-            votesCount = votesCount + choices.size,
-            votersCount = votersCount?.plus(1),
-            voted = true,
-        )
-    }
-
-    fun toNewPoll(creationDate: Date) = NewPoll(
-        options.map { it.title },
-        expiresAt?.let {
-            ((it.time - creationDate.time) / 1000).toInt() + 1
-        } ?: 3600,
-        multiple,
-    )
-
     fun asModel() = app.pachli.core.model.Poll(
         id = id,
         expiresAt = expiresAt,
