+++ feature/suggestions/src/main/kotlin/app/pachli/feature/suggestions/SuggestionsFragment.kt
@@ -44,13 +44,13 @@ import app.pachli.core.common.extensions.show
 import app.pachli.core.common.extensions.throttleFirst
 import app.pachli.core.common.extensions.viewBinding
 import app.pachli.core.common.util.unsafeLazy
-import app.pachli.core.data.model.SuggestionSources
-import app.pachli.core.data.model.SuggestionSources.FEATURED
-import app.pachli.core.data.model.SuggestionSources.FRIENDS_OF_FRIENDS
-import app.pachli.core.data.model.SuggestionSources.MOST_FOLLOWED
-import app.pachli.core.data.model.SuggestionSources.MOST_INTERACTIONS
-import app.pachli.core.data.model.SuggestionSources.SIMILAR_TO_RECENTLY_FOLLOWED
-import app.pachli.core.data.model.SuggestionSources.UNKNOWN
+import app.pachli.core.model.SuggestionSources
+import app.pachli.core.model.SuggestionSources.FEATURED
+import app.pachli.core.model.SuggestionSources.FRIENDS_OF_FRIENDS
+import app.pachli.core.model.SuggestionSources.MOST_FOLLOWED
+import app.pachli.core.model.SuggestionSources.MOST_INTERACTIONS
+import app.pachli.core.model.SuggestionSources.SIMILAR_TO_RECENTLY_FOLLOWED
+import app.pachli.core.model.SuggestionSources.UNKNOWN
 import app.pachli.core.navigation.AccountActivityIntent
 import app.pachli.core.navigation.TimelineActivityIntent
 import app.pachli.core.ui.BackgroundMessage
